+++ app/src/main/java/com/example/user/biblejournal/core/Test.java
@@ -1,0 +1,4 @@
+package com.example.user.biblejournal.core;
+
+public class Test {
+}
