+++ app/src/main/java/com/example/user/biblejournal/main/Test.java
@@ -1,0 +1,4 @@
+package com.example.user.biblejournal.main;
+
+public class Test {
+}
