--- app/src/main/java/com/example/user/biblejournal/model/database/bible/BookNameDao.kt
@@ -1,10 +1,0 @@
-package com.example.user.biblejournal.model.database.bible
-
-import androidx.room.Dao
-import androidx.room.Query
-
-@Dao
-interface BookNameDao {
-    @Query("SELECT id, name FROM bible_tagalog_books")
-    fun getAllBookNames() : List<BookNameEntity>
-}
