--- app/src/main/java/com/example/user/biblejournal/model/database/bible/MaxVerseEntity.kt
@@ -1,17 +1,0 @@
-package com.example.user.biblejournal.model.database.bible
-
-import androidx.annotation.NonNull
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "max_verse")
-data class MaxVerseEntity(
-        @NonNull
-        @PrimaryKey
-        var book : Int,
-
-        @NonNull
-        @ColumnInfo(name = "max_verses")
-        var maxVerses : String
-)
