+++ app/src/main/java/com/example/user/biblejournal/reader/Test.java
@@ -1,0 +1,4 @@
+package com.example.user.biblejournal.reader;
+
+public class Test {
+}
