+++ app/src/main/res/layout/fragment_note_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#272822"
-    tools:context=".view.notelist.NoteListFragment">
+    tools:context=".home.view.HomeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/note_list_recycler_view"
